+++ sudoku.cpp
@@ -1,5 +1,3 @@
-
-#include <windows.h>
 #include <algorithm>
 
 #include "data.h"
@@ -243,70 +241,6 @@ bool generate_board(int mod) {
     }
 }
 
-char get_num(int num) {
-    if (num == 0)
-        return '$';
-    else if (num > 0 && num < 10)
-        return num + '0';
-    else if (num > 9 && num < 17)
-        return num - 10 + 'A';
-    else {
-        //std::cout <<"【" << num<<"】";
-        return 'X';
-    }
-}
-
-void draw_board(const std::vector<std::vector<int>>& board) {
-    HANDLE hConsole = GetStdHandle(STD_OUTPUT_HANDLE);
-    // 字体颜色为红色
-    SetConsoleTextAttribute(hConsole, FOREGROUND_BLUE|FOREGROUND_GREEN);
-
-    int block = params.l;
-    int length = params.l * params.l;
-    int char_length = 2 * (1 + block + length) - 1;
-    for (int i = 0; i < char_length; i++) {
-        std::cout << "*";
-    }
-    std::cout << std::endl;
-
-    for (int i = 0; i < char_length; i++) {
-        // i % (length -1)：每个宫需要length-2个横线，所以每隔length-1个横线就需要添加一个加号
-        if (i == 0 || i % ( 2 * block + 2)==0) 
-            std::cout << "+";
-        else
-            std::cout << "-";
-    }
-    std::cout << std::endl;
-
-
-    for (int r = 0; r < length; r++) {
-
-        std::cout << "| ";
-        for (int c = 0; c < length; c++) {
-            std::cout << get_num(board[r][c]) << " ";
-            if ((c+1)%block==0)std::cout << "| ";
-        }
-        std::cout << std::endl;
-
-        // 打印横线
-        if ((r+1) % block == 0) {
-            for (int i = 0; i < char_length; i++) {
-                if (i == 0 || i % (2 * block + 2) == 0)
-                    std::cout << "+";
-                else
-                    std::cout << "-";
-            }
-            std::cout << std::endl;
-        }
-    }
-    for (int i = 0; i < char_length; i++) {
-        std::cout << "*";
-    }
-    std::cout << std::endl;
-    // 恢复默认的字体颜色
-    SetConsoleTextAttribute(hConsole, FOREGROUND_RED | FOREGROUND_GREEN | FOREGROUND_BLUE);
-}
-
 
 // 检查某个数字是否能填入指定的行、列和小九宫格
 bool is_valid(const std::vector<std::vector<int>>& board, int row, int col, int num) {
@@ -398,7 +332,7 @@ bool dig_hole(std::vector<std::vector<int>>& board) {
     std::uniform_int_distribution<> dis(1, max); // 假设生成 1 到 max 之间的随机数，用来选取x和y作为挖空的下标
     int rl = params.rl, rr = params.rr;
     int hole_cnt = 0; // 空计数
-    int x, y, z;
+    int x, y;
     while (hole_cnt <= rr){
         x = dis(gen);
         y = dis(gen);
